add tests of readme generator
