+++ package.json
@@ -5,7 +5,8 @@
   "main": "lib/textlint-rule-preset-track.js",
   "scripts": {
     "docs": "node script/generate_readme.js",
-    "test": "mocha test/ && (cd example && npm install && npm test)"
+    "docs:test": "mocha script/test/",
+    "test": "mocha test/ && npm run docs:test && (cd example && npm install && npm test)"
   },
   "repository": {
     "type": "git",
